Add files via upload
Changed how files are selected. While there may be a slight performance increase with large song libraries (1000+), the main purpose was to create a past song remembrance feature. Any new song played will be added to a "no play" list, where the size of the list is 33% of the current library. Once the list is full, any new song played will kick the last song off the list (The song that's been on the list the longest).